requirements.txt
PyQt6
face_recognition
Pillow
dlib